redefine widgets
